src/requirements.txt
python-dotenv
aiohttp
teams-ai>=1.4.0,<2.0.0
gunicorn
